+++ src/Fulma.Extensions.Wikiki.Switch/RELEASE_NOTES.md
@@ -1,3 +1,5 @@
 ### 1.0.0
 
 * Release stable version after Fulma.Extensions split
+* Use `LabelProps` and `InputProps` instead of `Props`
+* Add a feedback if `Switch.Id` isn't setted